api/requirements.txt
langchain>=0.1.20
langchain-community>=0.0.45
langchain-huggingface>=0.1.2
langchain-groq>=0.1.0
groq>=0.3.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
email-validator>=2.0.0



